Iri generator sometimes generates iri of entity already in db Fixes #855
